# docs/api_spec.yaml
openapi: 3.0.0
info:
  title: API de Autenticación
  version: "1.0.0"
  description: "Especificación de la API para registro, login y gestión de perfiles de usuario."
paths:
...
  /api/auth/refresh/:
    post:
      summary: "Refresca el token de acceso"
      operationId: refreshToken
      tags:
        - Autenticación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: "Token de acceso refrescado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: "Token de refresco inválido o expirado"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...